refactor: add css modules, decomposite app
