--- src/App.css
@@ -1,159 +1,0 @@
-body {
-  margin: 0;
-  box-sizing: border-box;
-  background-color: rgb(7, 6, 6);
-  text-align: center;
-}
-
-h1 {
-  color: #fff;
-}
-
-.app {
-  display: flex;
-  flex-direction: column;
-  height: 100vh;
-}
-
-.error-page {
-  color: white;
-  font-size: 1.5rem;
-  margin-top: 100px;
-}
-
-.error-page i {
-  font-size: 3rem;
-}
-
-.loader {
-  width: 48px;
-  height: 48px;
-  border: 5px solid #fff;
-  border-bottom-color: transparent;
-  border-radius: 50%;
-  display: inline-block;
-  box-sizing: border-box;
-  animation: rotation 1s linear infinite;
-}
-
-@keyframes rotation {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
-
-hr {
-  width: 100%;
-  color: black;
-}
-
-button {
-  width: 100px;
-  height: 40px;
-  border-radius: 5px;
-  background-color: rgba(0, 0, 0, 0.8);
-  color: #fff;
-  border: solid 1px #fff;
-  cursor: pointer;
-}
-
-.btn_page {
-  height: 40px;
-  border-radius: 5px;
-  background-color: rgba(0, 0, 0, 0.8);
-  color: #fff;
-  border: solid 1px #fff;
-  cursor: pointer;
-  padding: 5px 10px;
-  text-decoration: none;
-  margin-left: 8px;
-}
-
-.btn_page.active {
-  color: rgba(0, 0, 0, 0.8);
-  background-color: #fff;
-}
-
-.btn_page:hover,
-button:hover {
-  box-shadow: 2px 2px 5px 5px rgb(235, 220, 16);
-}
-
-input {
-  width: 240px;
-  height: 40px;
-  border-radius: 5px;
-  background-color: rgba(0, 0, 0, 0.8);
-  color: #fff;
-  border: solid 1px #fff;
-  cursor: text;
-  padding-left: 5px;
-}
-
-input::placeholder {
-  color: #fff;
-}
-
-.top-section {
-  padding: 20px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 20px;
-  height: 100px;
-}
-
-.main-content {
-  display: flex;
-  flex-direction: row;
-  flex: 2;
-}
-
-.pagination {
-  margin-top: 50px;
-  height: 80px;
-}
-
-.results-section {
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  height: calc(100vh - 230px);
-  gap: 20px;
-  flex-grow: 1;
-  padding: 20px;
-  overflow-y: auto;
-  color: #fff;
-}
-
-.detailed-card {
-  height: calc(100vh - 230px);
-  color: rgba(0, 0, 0, 0.8);
-  background-color: #fff;
-  padding: 20px;
-  width: 35%;
-}
-
-.detailed-card p {
-  margin: 5px;
-}
-
-.loader-container {
-  height: calc(100vh - 180px);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.result-item {
-  padding: 20px;
-  border: solid 1px #fff;
-  border-radius: 10px;
-  cursor: pointer;
-}
-
-.error {
-  color: red;
-}
