+++ src/components/Card/Card.tsx
@@ -1,9 +1,10 @@
-import { CardProps } from '../interfaces/interfaces';
+import { CardProps } from '../../interfaces/interfaces';
+import styles from './Card.module.css';
 
 function Card({ index, item, onClick }: CardProps) {
   return (
     <div
-      className='result-item'
+      className={styles.resultItem}
       data-testid='result-item'
       key={index}
       onClick={onClick}
