+++ src/components/CardList/CardList.tsx
@@ -1,5 +1,6 @@
-import { ResultsSectionProps } from '../interfaces/interfaces';
-import Card from './Card';
+import { ResultsSectionProps } from '../../interfaces/interfaces';
+import Card from '../Card/Card';
+import styles from './CardList.module.css';
 
 function CardList({ results }: ResultsSectionProps) {
   if (!results) {
@@ -9,7 +10,7 @@ function CardList({ results }: ResultsSectionProps) {
   const listItems = results.map((result, index) => (
     <Card key={result.name} index={index} item={result}></Card>
   ));
-  return <div className='results'>{listItems}</div>;
+  return <div className={styles.results}>{listItems}</div>;
 }
 
 export default CardList;
