+++ src/components/ErrorPage/ErrorPage.tsx
@@ -1,11 +1,12 @@
 import { useRouteError } from 'react-router-dom';
+import styles from './ErrorPage.module.css';
 
 export default function ErrorPage() {
   const error = useRouteError();
   console.error(error);
 
   return (
-    <div className='error-page'>
+    <div className={styles.errorPage}>
       <h1>Oops!</h1>
       <p>Sorry, an unexpected error has occurred.</p>
       <p>
