+++ src/components/Pagination/Pagination.module.css
@@ -1,0 +1,24 @@
+.btnPage {
+  height: 40px;
+  border-radius: 5px;
+  background-color: rgba(0, 0, 0, 0.8);
+  color: #fff;
+  border: solid 1px #fff;
+  cursor: pointer;
+  padding: 5px 10px;
+  text-decoration: none;
+  margin-left: 8px;
+}
+
+.btnPage.active {
+  color: rgba(0, 0, 0, 0.8);
+  background-color: #fff;
+}
+
+.btnPage:hover {
+  box-shadow: 2px 2px 5px 5px rgb(235, 220, 16);
+}
+
+.pagination {
+  height: 80px;
+}
