+++ src/components/SearchSection/SearchSection.module.css
@@ -1,0 +1,8 @@
+.searchSection {
+  padding: 20px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 20px;
+  height: 80px;
+}
