--- src/components/UI/Loader.tsx
@@ -1,9 +1,0 @@
-function Loader() {
-  return (
-    <div data-testid='loader-container' className='loader-container'>
-      <div data-testid='loader' className='loader'></div>
-    </div>
-  );
-}
-
-export default Loader;
