+++ src/components/UI/SearchInput/SearchInput.module.css
@@ -1,0 +1,14 @@
+.searchInput {
+  width: 240px;
+  height: 40px;
+  border-radius: 5px;
+  background-color: rgba(0, 0, 0, 0.8);
+  color: #fff;
+  border: solid 1px #fff;
+  cursor: text;
+  padding-left: 5px;
+}
+
+.searchInput::placeholder {
+  color: #fff;
+}
