+++ src/helpers/CardLoader.tsx
@@ -4,6 +4,6 @@ import { getPersonById } from '../helpers/api';
 export const CardLoader = async ({ params }: LoaderFunctionArgs) => {
   const { cardId } = params;
   if (!cardId) throw new Error('Card ID parameter is missing');
-  const data = await getPersonById(cardId.slice(-1));
+  const data = await getPersonById(cardId);
   return data;
 };
