+++ src/hooks/useSearchQuery.ts
@@ -8,8 +8,12 @@ export const useSearchQuery = (
     const localData = localStorage.getItem(key);
     return localData || defData;
   });
+
   useEffect(() => {
-    localStorage.setItem(key, searchQuery);
+    return () => {
+      localStorage.setItem(key, searchQuery);
+    };
   }, [key, searchQuery]);
+
   return [searchQuery, setSearchQuery];
 };
