+++ src/index.css
@@ -1,0 +1,19 @@
+body {
+  margin: 0;
+  box-sizing: border-box;
+  background-color: rgb(7, 6, 6);
+  text-align: center;
+}
+
+h1 {
+  color: #fff;
+}
+
+hr {
+  width: calc(100% - 40px);
+  color: black;
+}
+
+.error {
+  color: red;
+}
