+++ src/tests/CardTest.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen, fireEvent } from '@testing-library/react';
 import { describe, it, expect } from 'vitest';
-import Card from '../components/Card';
+import Card from '../components/Card/Card';
 import { mockResults } from './mock';
 
 const mockOnClick = vi.fn();
